faqchatbot/requirements.txt
# Core
django>=5.0,<6.0
djangorestframework>=3.15
django-environ>=0.11.2
gunicorn>=21.2.0
psycopg2-binary>=2.9.9   # PostgreSQL adapter

# Authentication & Security
django-allauth>=0.63.3
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.4.0

# Utilities
django-extensions>=3.2.3
pillow>=10.4.0            # For image uploads
requests>=2.32.3
python-decouple>=3.8
whitenoise>=6.6.0         # For static files in production

# Database Tools
dj-database-url>=2.2.0
sqlparse>=0.5.1


pytest>=8.3.3
pytest-django>=4.8.0

# Environment Variables
python-dotenv>=1.0.0
langchain>=0.1.0


